Renamed sudo to passthrough and moved functions to utils
