+++ crates/nu-cli/src/completions/completer.rs
@@ -9,6 +9,7 @@ use nu_protocol::{
     engine::{EngineState, Stack, StateWorkingSet},
     BlockId, PipelineData, Span, Value,
 };
+use nu_utils::utils::is_passthrough_command;
 use reedline::{Completer as ReedlineCompleter, Suggestion};
 use std::str;
 use std::sync::Arc;
@@ -141,11 +142,9 @@ impl NuCompleter {
                         let span_offset: usize = alias_offset.iter().sum();
                         let mut spans: Vec<String> = vec![];
 
+                        let is_passthrough_command =
+                            is_passthrough_command(self.engine_state.get_file_contents());
                         for (flat_idx, flat) in flattened.iter().enumerate() {
-                            let is_sudo_command = spans
-                                .first()
-                                .filter(|content| *content == &String::from("sudo"))
-                                .is_some();
                             // Read the current spam to string
                             let current_span = working_set.get_span_contents(flat.0).to_vec();
                             let current_span_str = String::from_utf8_lossy(&current_span);
@@ -240,7 +239,7 @@ impl NuCompleter {
                                 }
 
                                 // specially check if it is currently empty - always complete commands
-                                if (is_sudo_command && flat_idx == 1)
+                                if (is_passthrough_command && flat_idx == 1)
                                     || (flat_idx == 0
                                         && working_set.get_span_contents(new_span).is_empty())
                                 {
@@ -263,7 +262,7 @@ impl NuCompleter {
                                 }
 
                                 // Completions that depends on the previous expression (e.g: use, source-env)
-                                if (is_sudo_command && flat_idx > 1) || flat_idx > 0 {
+                                if (is_passthrough_command && flat_idx > 1) || flat_idx > 0 {
                                     if let Some(previous_expr) = flattened.get(flat_idx - 1) {
                                         // Read the content for the previous expression
                                         let prev_expr_str =
