+++ crates/nu-utils/src/utils.rs
@@ -93,3 +93,50 @@ pub fn perf(
         );
     }
 }
+
+pub fn find_non_whitespace_index(contents: &[u8], start: usize) -> usize {
+    contents[start..]
+        .iter()
+        .take_while(|x| x.is_ascii_whitespace())
+        .count()
+        + start
+}
+
+pub fn is_passthrough_command(working_set_file_contents: Vec<(Vec<u8>, usize, usize)>) -> bool {
+    for (contents, _, _) in working_set_file_contents {
+        let last_pipe_pos_rev = contents.iter().rev().position(|x| x == &b'|');
+        let last_pipe_pos = last_pipe_pos_rev.map(|x| contents.len() - x).unwrap_or(0);
+
+        let cur_pos = find_non_whitespace_index(&contents, last_pipe_pos);
+
+        if contents[cur_pos..].starts_with(b"sudo ") {
+            return true;
+        }
+    }
+    false
+}
+
+#[cfg(test)]
+mod utils_tests {
+    use super::*;
+
+    #[test]
+    fn test_find_non_whitespace_index() {
+        let commands = vec![
+            ("    hello", 4),
+            ("sudo ", 0),
+            (" 	sudo ", 2),
+            ("	 sudo ", 2),
+            ("	hello ", 1),
+            ("	  hello ", 3),
+            ("    hello | sudo ", 4),
+            ("     sudo|sudo", 5),
+            ("sudo | sudo ", 0),
+            ("	hello sud", 1),
+        ];
+        for (idx, ele) in commands.iter().enumerate() {
+            let index = find_non_whitespace_index(&Vec::from(ele.0.as_bytes()), 0);
+            assert_eq!(index, ele.1, "Failed on index {}", idx);
+        }
+    }
+}
